snippet bridge moved to runtime namespace
